+++ ios/ReactNativeSynerise/Modules/RNTracker.m
@@ -75,6 +75,11 @@ RCT_EXPORT_MODULE();
                         [builder setObject:((NSDictionary *)paramValue) forKey:paramKey];
                         continue;
                     }
+
+                    if ([paramValue isKindOfClass:[NSArray class]] == YES) {
+                        [builder setObject:((NSArray *)paramValue) forKey:paramKey];
+                        continue;
+                    }
                 }
             }
         }];