+++ lib/classes/events/recommendation/RecommendationEvent.d.ts
@@ -1,7 +1,4 @@
 import { CustomEvent } from '../CustomEvent';
 declare class RecommendationEvent extends CustomEvent {
-    constructor(label: string, action: string, productId: string, name: string, campaignId: string, campaignHash: string, parameters?: object);
-    setCategory(category: string): void;
-    setUrl(url: string): void;
 }
 export { RecommendationEvent };