+++ lib/classes/events/recommendation/RecommendationSeenEvent.d.ts
@@ -1,5 +1,7 @@
 import { RecommendationEvent } from './RecommendationEvent';
 declare class RecommendationSeenEvent extends RecommendationEvent {
-    constructor(label: string, productId: string, name: string, campaignId: string, campaignHash: string, parameters?: object);
+    constructor(label: string, productId: string, productName: string, campaignId: string, campaignHash: string, parameters?: object);
+    setCategory(category: string): void;
+    setUrl(url: string): void;
 }
 export { RecommendationSeenEvent };